Test the send/recieve of discriminant

+++ src/common/Protocol.java
@@ -4,7 +4,7 @@ public abstract class Protocol {
 	
 	static final public int PORT = 12345;
 	
-	static final public int QUERY_TEST = 0x0001;
+	static final public byte QUERY_TEST = 0x0001;
 	
 
 }
